Refactor morphism package
